% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.rmse}
\alias{.rmse}
\title{RMSE between two matrices}
\usage{
.rmse(A, B)
}
\arguments{
\item{A, B}{matrices of the same shape.}
}
\value{
non-negative scalar.
}
\description{
RMSE between two matrices
}
\keyword{internal}
