% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.standardize}
\alias{.standardize}
\title{Column standardization (mean 0, sd 1) returning attributes for predict}
\usage{
.standardize(X)
}
\arguments{
\item{X}{numeric matrix \eqn{N \times P}.}
}
\description{
Column standardization (mean 0, sd 1) returning attributes for predict
}
\keyword{internal}
