% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{perplexity}
\alias{perplexity}
\title{Perplexity of a fitted mixture model on a count matrix}
\usage{
perplexity(W, Theta, Phi)
}
\arguments{
\item{W}{observed count matrix \eqn{N \times V}.}

\item{Theta}{mixture proportions \eqn{N \times K}.}

\item{Phi}{component-category distributions \eqn{K \times V}.}
}
\value{
positive scalar.
}
\description{
\eqn{\mathrm{Perplexity} = \exp\{ -\sum_{i,v} w_{iv} \log
(\Theta \Phi)_{iv} / \sum_{i,v} w_{iv} \}}.
}
