% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.gscamm}
\alias{plot.gscamm}
\title{Default plot of GSCA-MM diagnostics}
\usage{
\method{plot}{gscamm}(x, ...)
}
\arguments{
\item{x}{a \code{gscamm} fit.}

\item{...}{passed to \code{plot}.}
}
\description{
Plots perplexity and the maximum parameter change per iteration.
}
