% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.gscamm_effects}
\alias{plot.gscamm_effects}
\title{Plot ALR coefficients with confidence intervals}
\usage{
\method{plot}{gscamm_effects}(x, drop_intercept = TRUE, sig_only = FALSE, ...)
}
\arguments{
\item{x}{a \code{gscamm_effects} object.}

\item{drop_intercept}{logical, drop intercept rows (default TRUE).}

\item{sig_only}{logical, restrict to significant adjusted p-values
below 0.05 (default FALSE).}

\item{...}{unused.}
}
\description{
Plot ALR coefficients with confidence intervals
}
